requirements.txt
aiocron==1.8
discord.py==1.7.3
dnspython==2.2.0
pymongo==4.0.1
pytz==2021.3
requests==2.27.1
